Fixes after working on review issues

+++ watermarks_implementation/src/DCTWatermark.h
@@ -20,21 +20,21 @@ public:
 		MIDDLE_FREQUENCY
 	};
 
-	DCTWatermark(double alpha = -1, double detectThreshold = -1, int skipCount = -1, const std::vector<int> reference = std::vector<int>(), WatermarkType type = DEFAULT, Detector::Type detectorType = Detector::CORRELATION_COEFFICIENT, EmbeddingType embeddingType = EmbeddingType::SKIP_FIRST);
+	DCTWatermark(double alpha = -1, double detectThreshold = -1, int skipCount = -1, const std::vector<double> reference = std::vector<double>(), WatermarkType type = DEFAULT, Detector::Type detectorType = Detector::CORRELATION_COEFFICIENT, EmbeddingType embeddingType = EmbeddingType::SKIP_FIRST);
 	void setAlpha(double alpha);
 	void setSkipCount(int skipCount);
 	void setDetectThreshold(double detectThreshold);
-	void setReference(const std::vector<int>& reference);
+	void setReference(const std::vector<double>& reference);
 
 	bool embed(std::shared_ptr<VideoFrame> pframe, bool value);
 	DetectResult detect(const std::shared_ptr<VideoFrame> pframe);
 	bool getCorrelation(const std::shared_ptr<VideoFrame> pframe, double& correlation);
 	bool getAverageDCT(const std::shared_ptr<VideoFrame> pframe, double& averageDCT);
 
-	static std::vector<int> createReference(const std::shared_ptr<VideoFrame> pframe, int maxValue, int skipCount, bool BarniVersion = false, EmbeddingType embeddingType = EmbeddingType::SKIP_FIRST);
-	static std::vector<int> cropReference(const std::vector<int> &reference, int skipCount, std::size_t x, std::size_t y, std::size_t width, std::size_t height, std::size_t widthOrig, std::size_t heightOrig);
-	static void referenceStore(const std::vector<int>& data, const std::string& filename);
-	static std::vector<int> referenceLoad(const std::string& filename);
+	static std::vector<double> createReference(const std::shared_ptr<VideoFrame> pframe, int maxValue, int skipCount, bool BarniVersion = false, EmbeddingType embeddingType = EmbeddingType::SKIP_FIRST);
+	static std::vector<double> cropReference(const std::vector<double> &reference, int skipCount, std::size_t x, std::size_t y, std::size_t width, std::size_t height, std::size_t widthOrig, std::size_t heightOrig);
+	static void referenceStore(const std::vector<double>& data, const std::string& filename);
+	static std::vector<double> referenceLoad(const std::string& filename);
 
 	static const int BLOCK_SIZE = 8;
 private:
@@ -44,7 +44,7 @@ private:
 	double                       m_alpha;
 	double                       m_detectThreshold;
 	int                          m_skipCount;
-	std::vector<int>             m_reference;
+	std::vector<double>          m_reference;
 	Detector::Type               m_detectorType;
 	WatermarkType                m_watermarkType;
 	DCTWatermark::EmbeddingType  m_embeddingType;
